backend/requirements.txt
Django==5.2.1
djangorestframework==3.14.0
django-filter==23.2
django-cors-headers==4.0.0
djoser==2.2.0
Pillow==9.5.0
python-dotenv==1.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
drf-spectacular==0.26.2